+++ src/Styling/navbar/Navbar.css
@@ -1,0 +1,33 @@
+/* Styling/navbar/Navbar.css */
+
+.navbar {
+    background-color: var(--primary-color);
+    color: white;
+    padding: 10px 20px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    box-shadow: var(--box-shadow);
+  }
+  
+  .navbar h1 {
+    margin: 0;
+    font-size: 24px;
+  }
+  
+  .nav-links {
+    display: flex;
+  }
+  
+  .nav-links a {
+    color: white;
+    margin-left: 20px;
+    text-decoration: none;
+    font-size: 16px;
+    transition: text-decoration 0.3s ease;
+  }
+  
+  .nav-links a:hover {
+    text-decoration: underline;
+  }
+
