+++ src/Styling/sidebar/sidebar.css
@@ -1,12 +1,17 @@
+/* Styling/sidebar/Sidebar.css */
+
 .sidebar {
     width: 200px;
-    background-color: #f4f4f4;
+    background-color: var(--background-color);
     padding: 20px;
-    box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1);
+    box-shadow: var(--box-shadow);
+    border-right: 1px solid var(--border-color);
   }
   
   .sidebar h3 {
     margin-top: 0;
+    color: var(--text-color);
+    font-size: 18px;
   }
   
   .sidebar ul {
@@ -20,10 +25,12 @@
   
   .sidebar ul li a {
     text-decoration: none;
-    color: #333;
+    color: var(--text-color);
+    transition: color 0.3s ease;
+    font-size: 16px;
   }
   
   .sidebar ul li a:hover {
-    color: #007BFF;
+    color: var(--primary-color);
   }
